charm fixes and loot table adjustments (epic charms, crystal skull)

+++ src/main/java/com/someguyssoftware/treasure2/item/charm/HarvestingCharm.java
@@ -80,7 +80,7 @@ public class HarvestingCharm extends Charm {
 	public void addInformation(ItemStack stack, World world, List<String> tooltip, ITooltipFlag flag, ICharmData data) {
 		TextFormatting color = TextFormatting.GREEN;
 		tooltip.add("  " + color + getLabel(data));
-		tooltip.add(" " + TextFormatting.GRAY + "" + TextFormatting.ITALIC + I18n.translateToLocalFormatted("tooltip.charm.harvest_rate", getMaxPercent()));
+		tooltip.add(" " + TextFormatting.GRAY + "" + TextFormatting.ITALIC + I18n.translateToLocalFormatted("tooltip.charm.harvest_rate", Math.toIntExact((long) getMaxPercent())));
 	}
 
 }
